support loading HF BERT checkpoints to our own TFM
this change makes it possible to load huggingface
transformer checkpoints (themselves converted from bert TF)
into our own TLMs.

to support this, we add a BERT embeddings which implements
token embeddings, and support T2T-style layer norms
