+++ layers/eight_mile/pytorch/layers.py
@@ -2754,8 +2754,11 @@ class TransformerEncoder(nn.Module):
         d_k: Optional[int] = None,
         rpr_k: Optional[int] = None,
         ffn_pdrop: Optional[float] = 0.0,
+        layer_norms_after=False
     ):
         super().__init__()
+        # to properly execute BERT models, we have to follow T2T and do layer norms after
+        self.layer_norms_after = layer_norms_after
         self.d_model = d_model
         self.d_ff = d_ff if d_ff is not None else 4 * d_model
         if rpr_k is not None:
@@ -2768,6 +2771,9 @@ class TransformerEncoder(nn.Module):
             nn.Dropout(ffn_pdrop),
             Dense(self.d_ff, self.d_model),
         )
+        # Slightly late for a name change
+        # LN1 = ln_x
+        # LN2 = ln_attn_output
         self.ln1 = nn.LayerNorm(self.d_model, eps=1e-6)
         self.ln2 = nn.LayerNorm(self.d_model, eps=1e-6)
         self.dropout = nn.Dropout(pdrop)
@@ -2779,14 +2785,14 @@ class TransformerEncoder(nn.Module):
         """
         x, mask = inputs
 
-        x = self.ln1(x)
+        if not self.layer_norms_after:
+            x = self.ln1(x)
         h = self.self_attn((x, x, x, mask))
-
         x = x + self.dropout(h)
-
         x = self.ln2(x)
         x = x + self.dropout(self.ffn(x))
-        ##x = self.ln1(x)
+        if self.layer_norms_after:
+            x = self.ln1(x)
         return x
 
 
@@ -2853,11 +2859,12 @@ class TransformerEncoderStack(nn.Module):
         d_k: Optional[int] = None,
         rpr_k: Optional[Union[int, List[int]]] = None,
         ffn_pdrop: Optional[float] = 0.0,
+        layer_norms_after = False,
         **kwargs,
     ):
         super().__init__()
         self.encoders = nn.ModuleList()
-        self.ln = nn.LayerNorm(d_model, eps=1e-6)
+        self.ln = nn.Identity() if layer_norms_after else nn.LayerNorm(d_model, eps=1e-6)
         self.output_dim = d_model
 
         if not is_sequence(rpr_k):
@@ -2866,7 +2873,8 @@ class TransformerEncoderStack(nn.Module):
         for i in range(layers):
             self.encoders.append(
                 TransformerEncoder(
-                    num_heads, d_model, pdrop, scale, activation, d_ff, d_k, rpr_k=rpr_k[i], ffn_pdrop=ffn_pdrop
+                    num_heads, d_model, pdrop, scale, activation, d_ff, d_k,
+                    rpr_k=rpr_k[i], ffn_pdrop=ffn_pdrop, layer_norms_after=layer_norms_after
                 )
             )
 
@@ -2890,6 +2898,7 @@ class TransformerEncoderStackWithLengths(TransformerEncoderStack):
         d_k: Optional[int] = None,
         rpr_k: Optional[Union[int, List[int]]] = None,
         input_sz: Optional[int] = None,
+        layer_norms_after = False,
         **kwargs,
     ):
         super().__init__(num_heads, d_model, pdrop, scale, layers, activation, d_ff, d_k, rpr_k)
@@ -2917,9 +2926,10 @@ class TransformerEncoderStackWithTimeMask(TransformerEncoderStack):
         d_k: Optional[int] = None,
         rpr_k: Optional[Union[int, List[int]]] = None,
         input_sz: Optional[int] = None,
+        layer_norms_after = False,
         **kwargs,
     ):
-        super().__init__(num_heads, d_model, pdrop, scale, layers, activation, d_ff, d_k, rpr_k)
+        super().__init__(num_heads, d_model, pdrop, scale, layers, activation, d_ff, d_k, rpr_k, layer_norms_after=layer_norms_after)
         self.proj = WithDropout(pytorch_linear(input_sz, d_model), pdrop)
 
     def forward(self, inputs: Tuple[torch.Tensor, torch.Tensor]) -> torch.Tensor:
